removed deprecated central classes ServicePlatform, AbstractView and others. added new mock objects for testing. added initial environment for plugin loading. please note this revision cannot be started due to serialisation issues.

+++ honeycrm_r134_gwt2.1m2/src/honeycrm/client/mvp/presenters/DetailPresenter.java
@@ -131,7 +131,7 @@ public class DetailPresenter implements Presenter {
 	
 	protected HashMap<String, Object> getDefaultPrefilledFields() {
 		final HashMap<String, Object> prefilledFields = new HashMap<String, Object>();
-		prefilledFields.put("assignedTo", (Long) User.getUserId());
+		prefilledFields.put("assignedTo", User.getUserId());
 		return prefilledFields;
 	}
